requirements.txt
txtai
streamlit
PyPDF2
youtube_transcript_api
google-generativeai
python-dotenv
pathlib